requirements/development.txt
-r base.txt
django-debug-toolbar==1.5
Werkzeug==0.11.11
